Added test files in order to keep original files
